[front] - feat (AB v2) refactor template form transform
