+++ front/components/agent_builder/AgentBuilder.tsx
@@ -14,8 +14,8 @@ import { submitAgentBuilderForm } from "@app/components/agent_builder/submitAgen
 import {
   getDefaultAgentFormData,
   transformAgentConfigurationToFormData,
+  transformTemplateToFormData,
 } from "@app/components/agent_builder/transformAgentConfiguration";
-import { AGENT_CREATIVITY_LEVEL_TEMPERATURES } from "@app/components/agent_builder/types";
 import { ConversationSidePanelProvider } from "@app/components/assistant/conversation/ConversationSidePanelContext";
 import { appLayoutBack } from "@app/components/sparkle/AppContentLayout";
 import { FormProvider } from "@app/components/sparkle/FormProvider";
@@ -57,42 +57,8 @@ export default function AgentBuilder({
       return transformAgentConfigurationToFormData(agentConfiguration);
     }
     
-    // If we have a template but no agent configuration, apply template data
     if (assistantTemplate) {
-      const defaultFormData = getDefaultAgentFormData(user);
-      const templateFormData = {
-        ...defaultFormData,
-        instructions:
-          assistantTemplate.presetInstructions ?? defaultFormData.instructions,
-        agentSettings: {
-          ...defaultFormData.agentSettings,
-          name: assistantTemplate.handle ?? defaultFormData.agentSettings.name,
-          description:
-            assistantTemplate.description ??
-            defaultFormData.agentSettings.description,
-          pictureUrl:
-            assistantTemplate.pictureUrl ??
-            defaultFormData.agentSettings.pictureUrl,
-        },
-        generationSettings: {
-          ...defaultFormData.generationSettings,
-          modelSettings: {
-            ...defaultFormData.generationSettings.modelSettings,
-            providerId:
-              assistantTemplate.presetProviderId ??
-              defaultFormData.generationSettings.modelSettings.providerId,
-            modelId:
-              assistantTemplate.presetModelId ??
-              defaultFormData.generationSettings.modelSettings.modelId,
-          },
-          temperature: assistantTemplate.presetTemperature
-            ? AGENT_CREATIVITY_LEVEL_TEMPERATURES[
-                assistantTemplate.presetTemperature
-              ]
-            : defaultFormData.generationSettings.temperature,
-        },
-      };
-      return templateFormData;
+      return transformTemplateToFormData(assistantTemplate, user);
     }
     
     return getDefaultAgentFormData(user);
